requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
qrcode[pil]==7.4.2
Pillow==10.0.1
python-socketio==5.9.0
eventlet==0.33.3
